Save coverage to a file
Save coverage order in $XDG_CONFIG_HOME/pango/coverage.order.

We compare the hashes of all fonts to identify when the
coverage order needs to be recomputed.

Rough numbers show that loading the coverage from a file
takes < 10ms, while computing it from scratch takes >400ms.
So this is clearly a win.